+++ src/lib/KeyLoaders/index.ts
@@ -1,0 +1,3 @@
+export { default as KeyLoader } from './KeyLoader';
+export { default as KeyknoxLoader } from './KeyknoxLoader';
+export { default as LocalLoader } from './LocalLoader';
